feat: change cra to vite
