+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_API_URL='https://panda-market-api.vercel.app/products'
+ViTE_API_URL='https://panda-market-api.vercel.app/products'
