+++ .prettierrc.json
@@ -1,5 +1,5 @@
 {
-  "printWidth": 80,
+  "printWidth": 100,
   "singleQuote": true,
   "jsxSingleQuote": true,
   "tabWidth": 2,
