+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["vite.config.ts", "tsconfig.json", "src"],
+  "ext": "ts,tsx,js,jsx,json",
+  "ignore": ["node_modules", ".git"],
+  "exec": "vite"
+}
