+++ src/features/item-list/hook/usePageSize.tsx
@@ -1,5 +1,5 @@
-import { DeviceType } from 'shared/store';
-import { useDeviceStore } from 'shared/store/useDeviceStore';
+import { DeviceType } from 'shared/lib';
+import { useDeviceType } from 'shared/store';
 
 // 상품 목록 타입 정의
 type ItemType = 'all' | 'best';
@@ -8,7 +8,7 @@ const DEFAULT_PAGE_SIZE: number = 10;
 
 // 페이지 사이즈 반환 커스텀 훅
 export function usePageSize(type: ItemType) {
-  const { deviceType } = useDeviceStore();
+  const deviceType = useDeviceType();
 
   const pageSizeMap: { [key in ItemType]: { [key in DeviceType]: number } } = {
     all: {
