+++ src/features/item-list/lib/config.ts
@@ -1,2 +1,2 @@
 // 환경변수 타입 정의
-export const BASE_URL: string | undefined = process.env.REACT_APP_API_URL;
+export const BASE_URL = import.meta.env.VITE_API_URL;
