+++ src/features/item-list/lib/index.ts
@@ -1,3 +1,5 @@
 export { BASE_URL } from './config';
+
 export type { GetItemsParams, Item, GetItemsResponse } from './api';
+
 export { getItems } from './getItems';
