+++ src/features/item-list/ui/DropDown.tsx
@@ -1,9 +1,61 @@
 import { useState } from 'react';
+
 import styled from 'styled-components';
+
 import ArrowDownIcon from 'public/images/ic_arrow_down.svg';
 import SortIcon from 'public/images/ic_sort.svg';
-import { DeviceType } from 'shared/store';
-import { useDeviceStore } from 'shared/store/useDeviceStore';
+
+import { GetItemsParams } from 'features/item-list/lib';
+
+import { useDeviceType } from 'shared/store';
+
+export function DropDown({ order, setOrder }) {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const deviceType = useDeviceType();
+
+  const isMobile = deviceType === 'mobile';
+  const buttonText = order === 'recent' ? '최신순' : '좋아요순';
+  const orderType = [
+    {
+      type: 'recent',
+      text: '최신순',
+    },
+    {
+      type: 'favorite',
+      text: '좋아요순',
+    },
+  ];
+
+  const handleToggle = () => setIsOpen(!isOpen);
+  const handleChangeOrder = (newOrder) => {
+    setOrder(newOrder);
+    setIsOpen(false);
+  };
+
+  return (
+    <>
+      <DropDownButton $deviceType={deviceType} onClick={handleToggle}>
+        {!isMobile && <SortByItem>{buttonText}</SortByItem>}
+        <DropDownIcon
+          $deviceType={deviceType}
+          src={isMobile ? SortIcon : ArrowDownIcon}
+          alt='ArrowDownIcon'
+        />
+      </DropDownButton>
+
+      {isOpen && (
+        <DropDownMenu>
+          {orderType.map((item) => (
+            <DropDownItem key={item.type} onClick={() => handleChangeOrder(item.type)}>
+              {item.text}
+            </DropDownItem>
+          ))}
+        </DropDownMenu>
+      )}
+    </>
+  );
+}
 
 const DropDownButton = styled.button`
   display: flex;
@@ -13,8 +65,7 @@ const DropDownButton = styled.button`
   border: 1px solid var(--gray200);
   background-color: #fff;
   font-size: 16px;
-  padding: ${({ $deviceType }) =>
-    $deviceType === 'mobile' ? '0 9px' : '0 20px'};
+  padding: ${({ $deviceType }) => ($deviceType === 'mobile' ? '0 9px' : '0 20px')};
 `;
 
 const DropDownIcon = styled.img`
@@ -52,54 +103,3 @@ const DropDownItem = styled.li`
   }
   cursor: pointer;
 `;
-
-function DropDown({ deviceType, order, setOrder }: DeviceType) {
-  const [isOpen, setIsOpen] = useState(false);
-  const { isMobile } = useDeviceStore();
-
-  const buttonText = order === 'recent' ? '최신순' : '좋아요순';
-  const orderType = [
-    {
-      type: 'recent',
-      text: '최신순',
-    },
-    {
-      type: 'favorite',
-      text: '좋아요순',
-    },
-  ];
-
-  const handleToggle = () => setIsOpen(!isOpen);
-  const handleChangeOrder = (newOrder) => {
-    setOrder(newOrder);
-    setIsOpen(false);
-  };
-
-  return (
-    <>
-      <DropDownButton $deviceType={deviceType} onClick={handleToggle}>
-        {!isMobile && <SortByItem>{buttonText}</SortByItem>}
-        <DropDownIcon
-          $deviceType={deviceType}
-          src={isMobile ? SortIcon : ArrowDownIcon}
-          alt='ArrowDownIcon'
-        />
-      </DropDownButton>
-
-      {isOpen && (
-        <DropDownMenu>
-          {orderType.map((item) => (
-            <DropDownItem
-              key={item.type}
-              onClick={() => handleChangeOrder(item.type)}
-            >
-              {item.text}
-            </DropDownItem>
-          ))}
-        </DropDownMenu>
-      )}
-    </>
-  );
-}
-
-export default DropDown;
