+++ src/main.tsx
@@ -1,9 +1,10 @@
 import { StrictMode } from 'react';
+
 import ReactDOM from 'react-dom/client';
+
 import App from 'app/App';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <App />
   </StrictMode>,
