+++ src/shared/lib/Device.ts
@@ -1,0 +1,2 @@
+// 디바이스 타입 정의
+export type DeviceType = 'mobile' | 'tablet' | 'desktop';
