+++ src/shared/lib/index.ts
@@ -1,1 +1,4 @@
+export type { DeviceType } from './Device';
+
 export { handleAxiosError } from './handleAxiosError';
+export { getDeviceType } from './getDeviceType';
