+++ src/widgets/item-list/index.ts
@@ -1,0 +1,1 @@
+export { ItemList } from './ItemList';
