+++ src/widgets/layout/Header.tsx
@@ -36,8 +36,7 @@ const Container = styled.div`
 const Wrapper = styled.div`
   display: flex;
   justify-content: space-between;
-  column-gap: ${({ $deviceType }) =>
-    $deviceType === 'mobile' ? '8px' : '39px;'};
+  column-gap: ${({ $deviceType }) => ($deviceType === 'mobile' ? '8px' : '39px;')};
 `;
 
 export function Header() {
