+++ src/widgets/layout/Layout.tsx
@@ -1,7 +1,9 @@
 import { Outlet } from 'react-router-dom';
+
 import { Helmet } from 'react-helmet';
-import { Header } from './Header';
+
 import { Footer } from './Footer';
+import { Header } from './Header';
 
 export function Layout() {
   return (
